Pulled contributions landing reducer out into shared reducer.
