+++ assets/pages/contributions-landing-us/contributionsLandingUSReducer.js
@@ -1,29 +1,8 @@
 // @flow
 
-// ----- Imports ----- //
+// // ----- Imports ----- //
 
-import { combineReducers } from 'redux';
-
-import { contributionSelectionReducerFor } from 'containerisableComponents/contributionSelection/contributionSelectionReducer';
-import { payPalContributionButtonReducerFor } from 'containerisableComponents/payPalContributionButton/payPalContributionButtonReducer';
-
-import type { State as ContributionSelectionState } from 'containerisableComponents/contributionSelection/contributionSelectionReducer';
-import type { State as PayPalState } from 'containerisableComponents/payPalContributionButton/payPalContributionButtonReducer';
-
-import type { CommonState } from 'helpers/page/page';
-
-
-// ----- Types ----- //
-
-type PageState = {
-  selection: ContributionSelectionState,
-  payPal: PayPalState,
-};
-
-export type State = {
-  common: CommonState,
-  page: PageState,
-};
+import { reducer } from 'reducers/contributions-landing/contributionsLandingReducer';
 
 
 // ----- Initial State ----- //
@@ -37,7 +16,4 @@ const initialStateOverrides = {
 
 // ----- Reducer ----- //
 
-export default combineReducers({
-  selection: contributionSelectionReducerFor('CONTRIBUTE_SECTION', initialStateOverrides),
-  payPal: payPalContributionButtonReducerFor('CONTRIBUTE_SECTION'),
-});
+export default reducer(initialStateOverrides);
